requirements.txt
Django>=4.0.3,<4.1.0
djangorestframework>=3.13.1,<3.14.0

flake8>=4.0.1,<4.1.0
psycopg2>=2.8.6,<2.9.0
drf-spectacular>=0.15.1,<0.16
Pillow>=8.2.0,<8.3.2